requirements.txt
dota2gsipy @ git+https://github.com/Daniel-EST/dota2gsipy.git
flask
google-genai
python-dotenv
